chore: 게시글 모달 관련 컴포넌트 폴더 이동
